compact with linux/cygwin
git-svn-id: https://cppe4x.googlecode.com/svn/trunk@34 a68ca59a-e7b1-11dd-9059-71fac19e68c6

+++ CppE4X/E4XIterator.cpp
@@ -1,5 +1,6 @@
 #include "E4X.h"
 #include "E4XCharSet.h"
+#include <cassert>
 
 namespace E4X
 {
@@ -245,7 +246,7 @@ namespace E4X
 		return getCell().toNumber();
 	}
 
-	__int64 E4XIterator::toBigNumber()
+	long long E4XIterator::toBigNumber()
 	{
 		return getCell().toBigNumber();
 	}
